+++ src/code42cli/cmds/high_risk_employee.py
@@ -20,6 +20,9 @@ from code42cli.cmds.shared import get_user_id
 from code42cli.file_readers import read_csv_arg
 from code42cli.options import format_option
 from code42cli.options import sdk_options
+from code42cli.util import deprecation_warning
+
+DEPRECATION_TEXT = "(DEPRECATED): Use `code42 watchlists` commands instead."
 
 
 def _get_filter_choices():
@@ -45,65 +48,79 @@ risk_tag_option = click.option(
 )
 
 
-@click.group(cls=OrderedGroup)
+@click.group(
+    cls=OrderedGroup,
+    help=f"{DEPRECATION_TEXT}\n\nAdd and remove employees from the High Risk Employees detection list.",
+)
 @sdk_options(hidden=True)
 def high_risk_employee(state):
-    """Add and remove employees from the High Risk Employees detection list."""
     pass
 
 
-@high_risk_employee.command("list")
+@high_risk_employee.command(
+    "list",
+    help=f"{DEPRECATION_TEXT}\n\nLists the employees on the High Risk Employee list.",
+)
 @sdk_options()
 @format_option
 @filter_option
 def _list(state, format, filter):
-    """Lists the employees on the High Risk Employee list."""
-
+    deprecation_warning(DEPRECATION_TEXT)
     employee_generator = _get_high_risk_employees(state.sdk, filter)
     list_employees(employee_generator, format)
 
 
-@high_risk_employee.command()
+@high_risk_employee.command(
+    help=f"{DEPRECATION_TEXT}\n\nAdd a user to the high risk employees detection list."
+)
 @cloud_alias_option
 @notes_option
 @risk_tag_option
 @username_arg
 @sdk_options()
 def add(state, username, cloud_alias, risk_tag, notes):
-    """Add a user to the high risk employees detection list."""
+    deprecation_warning(DEPRECATION_TEXT)
     _add_high_risk_employee(state.sdk, username, cloud_alias, risk_tag, notes)
 
 
-@high_risk_employee.command()
+@high_risk_employee.command(
+    help=f"{DEPRECATION_TEXT}\n\nRemove a user from the high risk employees detection list."
+)
 @username_arg
 @sdk_options()
 def remove(state, username):
-    """Remove a user from the high risk employees detection list."""
+    deprecation_warning(DEPRECATION_TEXT)
     _remove_high_risk_employee(state.sdk, username)
 
 
-@high_risk_employee.command()
+@high_risk_employee.command(
+    help=f"{DEPRECATION_TEXT}\n\nAssociates risk tags with a user."
+)
 @username_arg
 @risk_tag_option
 @sdk_options()
 def add_risk_tags(state, username, risk_tag):
-    """Associates risk tags with a user."""
+    deprecation_warning(DEPRECATION_TEXT)
     _add_risk_tags(state.sdk, username, risk_tag)
 
 
-@high_risk_employee.command()
+@high_risk_employee.command(
+    help=f"{DEPRECATION_TEXT}\n\nDisassociates risk tags from a user."
+)
 @username_arg
 @risk_tag_option
 @sdk_options()
 def remove_risk_tags(state, username, risk_tag):
-    """Disassociates risk tags from a user."""
+    deprecation_warning(DEPRECATION_TEXT)
     _remove_risk_tags(state.sdk, username, risk_tag)
 
 
-@high_risk_employee.group(cls=OrderedGroup)
+@high_risk_employee.group(
+    cls=OrderedGroup,
+    help=f"{DEPRECATION_TEXT}\n\nTools for executing high risk employee actions in bulk.",
+)
 @sdk_options(hidden=True)
 def bulk(state):
-    """Tools for executing high risk employee actions in bulk."""
     pass
 
 
@@ -125,12 +142,13 @@ bulk.add_command(high_risk_employee_generate_template)
 
 @bulk.command(
     name="add",
-    help="Bulk add users to the high risk employees detection list using a CSV file with "
-    f"format: {','.join(HIGH_RISK_EMPLOYEE_CSV_HEADERS)}.",
+    help=f"{DEPRECATION_TEXT}\n\nBulk add users to the high risk employees detection list using a "
+    f"CSV file with format: {','.join(HIGH_RISK_EMPLOYEE_CSV_HEADERS)}.",
 )
 @read_csv_arg(headers=HIGH_RISK_EMPLOYEE_CSV_HEADERS)
 @sdk_options()
 def bulk_add(state, csv_rows):
+    deprecation_warning(DEPRECATION_TEXT)
     sdk = state.sdk
 
     def handle_row(username, cloud_alias, risk_tag, notes):
@@ -145,11 +163,13 @@ def bulk_add(state, csv_rows):
 
 @bulk.command(
     name="remove",
-    help=f"Bulk remove users from the high risk employees detection list using a CSV file with format {','.join(REMOVE_EMPLOYEE_HEADERS)}.",
+    help=f"{DEPRECATION_TEXT}\n\nBulk remove users from the high risk employees detection list "
+    f"using a CSV file with format {','.join(REMOVE_EMPLOYEE_HEADERS)}.",
 )
 @read_csv_arg(headers=REMOVE_EMPLOYEE_HEADERS)
 @sdk_options()
 def bulk_remove(state, csv_rows):
+    deprecation_warning(DEPRECATION_TEXT)
     sdk = state.sdk
 
     def handle_row(username):
@@ -164,12 +184,13 @@ def bulk_remove(state, csv_rows):
 
 @bulk.command(
     name="add-risk-tags",
-    help=f"Adds risk tags to users in bulk using a CSV file with format: "
+    help=f"{DEPRECATION_TEXT}\n\nAdds risk tags to users in bulk using a CSV file with format: "
     f"{','.join(RISK_TAG_CSV_HEADERS)}.",
 )
 @read_csv_arg(headers=RISK_TAG_CSV_HEADERS)
 @sdk_options()
 def bulk_add_risk_tags(state, csv_rows):
+    deprecation_warning(DEPRECATION_TEXT)
     sdk = state.sdk
 
     def handle_row(username, tag):
@@ -184,12 +205,13 @@ def bulk_add_risk_tags(state, csv_rows):
 
 @bulk.command(
     name="remove-risk-tags",
-    help=f"Removes risk tags from users in bulk using a CSV file with format: "
-    f"{','.join(RISK_TAG_CSV_HEADERS)}.",
+    help=f"{DEPRECATION_TEXT}\n\nRemoves risk tags from users in bulk using a CSV file with "
+    f"format: {','.join(RISK_TAG_CSV_HEADERS)}.",
 )
 @read_csv_arg(headers=RISK_TAG_CSV_HEADERS)
 @sdk_options()
 def bulk_remove_risk_tags(state, csv_rows):
+    deprecation_warning(DEPRECATION_TEXT)
     sdk = state.sdk
 
     def handle_row(username, tag):