+++ src/code42cli/main.py
@@ -2,6 +2,7 @@ import os
 import signal
 import site
 import sys
+import warnings
 
 import click
 from click_plugins import with_plugins
@@ -24,7 +25,10 @@ from code42cli.cmds.securitydata import security_data
 from code42cli.cmds.shell import shell
 from code42cli.cmds.trustedactivities import trusted_activities
 from code42cli.cmds.users import users
+from code42cli.cmds.watchlists import watchlists
 from code42cli.options import sdk_options
+
+warnings.simplefilter("ignore", DeprecationWarning)
 
 
 # Handle KeyboardInterrupts by just exiting instead of printing out a stack
@@ -93,3 +97,4 @@ cli.add_command(security_data)
 cli.add_command(shell)
 cli.add_command(users)
 cli.add_command(trusted_activities)
+cli.add_command(watchlists)